Cloudwatch and Kinesis stats fetched
Everything works on a per day basis, hard coded to yesterday in the
main execute function. This seems the simplest solution for now. The
intention is this checker will be run in the early hours every
morning.

The count is sum of the IncomingRecords metric for Kenisis is fetched
for yesterday and all S3 event log files are downloaded and
counted. The two counts are then compared, with an exception being
thrown if they do not match.

Remaining is how we hook this up to alerting, which will probably need
some playing with logging and that exception.